Added SecondAttempt.thy
Once the error in our definition was found, we made adequate changes to the theory. However, we still found that the additivity of our derivative operator requires further conditions to be true. We explain them at the end of the file next to the "sorry" statements. Two directions are possible: add the conditions or modify the definition of addition of polynomials...